Render right_page onboarding template correctly
Currently, the template_content.py DeviceContent class returns `None`
if onboarding is not enabled for an OnboardingDevice object. Likewise,
if no OnboardingDevice object exists, the template continues trying to
access attributes for an OnboardingDevice object. In the first case,
template rendering will fail as an empty string is needed in order
to insert nothing into the rendered HTML template presented to the user.
In the second case, an AttributeError is raised as you can not access
attributes of a NoneType object.