Ensure device_data is thread safe.
This adds a lock on the Processor so that multiple tasks don't overwrite each other, also ensures that we are working off of the values in the dictionary at the point the Napalm tasks completed. This ensures that an alteration to device_data that happens after the loop is started will not effect the onboarding.