Close DB connections when exiting forked processes
When running the SSOTSyncNetworkData job with 10 devices, the DB connection count would increment by 20, which meant there were two places the code was forking. The first was with the Nornir calling netmiko_send_commands, and the second was with the CommandGetterProcessor in the Nornir setup. Wrapping both of these with the connections.close_call() would close those 20 db connections (10 in one and 10 in the other) after they were finished without breaking the job.